Create key _and_rooms_BST.java
This Java code defines a class Solution with methods to determine if it’s possible to visit all rooms in a given list of rooms. Each room has a list of keys to other rooms. Here’s a breakdown of the code:

canVisitAllRooms(List<List<Integer>> rooms): This is the main method that checks if all rooms can be visited.
It initializes an array vis to keep track of visited rooms.
It calls the dfs method starting from room 0.
After the dfs call, it checks if there are any rooms that haven’t been visited by iterating through the vis array. If it finds an unvisited room (false), it returns false.
If all rooms have been visited, it returns true.

dfs(List<List<Integer>> rooms, boolean[] vis, int cur): This is a helper method that uses Depth-First Search (DFS) to visit rooms.
It checks if the current room cur has already been visited. If so, it returns immediately to avoid revisiting.
It marks the current room as visited (vis[cur] = true).
It then iterates through all keys in the current room and recursively calls dfs on each room that can be accessed with those keys.